+++ Media Player SDK X (crossplatform)/Avalonia/Simple Video Player Avalonia/MainWindow.axaml.cs
@@ -247,6 +247,7 @@ namespace Simple_Video_Player_Avalonia
             Dispatcher.UIThread.InvokeAsync(() =>
             {
                 VideoView1.Refresh();
+                //ShowMessage("Playback complete.");
             });
         }